+++ R/dbl.R
@@ -1,9 +1,9 @@
 #' Check Dbl
 #' 
-#' Checks if x is a dbl (non-missing numeric scalar).
+#' Checks if x is a dbl (non-missing numeric scalar with no attributes including names).
 #'
 #' @param x The object to check.
-#' @param coerce A flag indicating whether to coerce a scalar integer to a dbl.
+#' @param coerce A flag indicating whether to coerce a scalar integer to a dbl and drop attributes including names.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
@@ -19,12 +19,12 @@ check_dbl <- function(x, coerce = FALSE,
                        error = TRUE) {
   x_name <- deparse_x_name(x_name)
 
-
   check_flag_internal(coerce)
   check_flag_internal(error)
 
-  if(coerce && is.numeric(x)) {
-    x <- as.numeric(x)     
+  if(coerce) {
+    attributes(x) <- NULL
+    if(is.integer(x)) x <- as.double(x)     
   }
   
   check_scalar(x, values = 1, x_name = x_name, error = error)