+++ R/flag.R
@@ -1,8 +1,9 @@
 #' Check Flag
 #' 
-#' Checks if x is a flag (non-missing logical scalar).
+#' Checks if x is a flag (non-missing logical scalar with no attributes including names).
 #'
 #' @param x The object to check.
+#' @param coerce A flag indicating whether to drop attributes including names.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
@@ -13,15 +14,18 @@
 #' check_flag(1, error = FALSE)
 #' check_flag(FALSE, error = FALSE)
 #' check_flag(c(FALSE, TRUE), error = FALSE)
-check_flag <- function(x,
+check_flag <- function(x, coerce = FALSE,
                        x_name = substitute(x),
                        error = TRUE) {
   x_name <- deparse_x_name(x_name)
 
+  check_flag_internal(coerce)
   check_flag_internal(error)
+  
+  if(coerce)
+    attributes(x) <- NULL
 
-  check_scalar(x, values = TRUE,
-               x_name = x_name, error = error)
+  check_scalar(x, values = TRUE, x_name = x_name, error = error)
 }
 
 #' Check Flag or NA
@@ -30,9 +34,7 @@ check_flag <- function(x,
 #' 
 #' Useful when using flag to pass one of three options.
 #'
-#' @param x The object to check.
-#' @param x_name A string of the name of the object.
-#' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
+#' @inheritParams check_flag
 #' @return An invisible copy of x (if it doesn't throw an error).
 #' @export
 #'
@@ -42,14 +44,16 @@ check_flag <- function(x,
 #' check_flag_na(NA)
 #' check_flag_na(NA_character_, error = FALSE)
 #' check_flag_na(c(NA, NA), error = FALSE)
-check_flag_na <- function(x,
+check_flag_na <- function(x, coerce = TRUE,
                        x_name = substitute(x),
                        error = TRUE) {
   x_name <- deparse_x_name(x_name)
 
+  check_flag_internal(coerce)
   check_flag_internal(error)
+  
+  if(coerce)
+    attributes(x) <- NULL
 
-  check_scalar(x, values = c(TRUE, NA),
-               x_name = x_name, error = error)
+  check_scalar(x, values = c(TRUE, NA), x_name = x_name, error = error)
 }
-