+++ R/probability.R
@@ -1,9 +1,9 @@
 #' Check Probability
 #' 
-#' Checks if x is a probability (non-missing double scalar between 0 and 1 inclusive or if \code{coerce = TRUE} an integer of 0 or 1).
+#' Checks if x is a probability (non-missing dbl between 0 and 1 inclusive with no attributes including names).
 #'
 #' @param x The object to check.
-#' @param coerce A flag indicating whether to coerce an integer to numeric.
+#' @param coerce A flag indicating whether to coerce an integer to a probability and drop attributes including names.
 #' @param x_name A string of the name of the object.
 #' @param error A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.
 #' @return An invisible copy of x (if it doesn't throw an error).
@@ -22,9 +22,10 @@ check_probability <- function(x, coerce = FALSE,
   check_flag_internal(coerce)
   check_flag_internal(error)
   
-  if(coerce && is.integer(x))
-    x <- as.numeric(x)
-
-  check_vector(x, values = c(0,1), length = 1,
-               x_name = x_name, error = error)
+  if(coerce) {
+    attributes(x) <- NULL
+    if(is.integer(x))
+      x <- as.numeric(x)
+  }
+  check_scalar(x, values = c(0,1), x_name = x_name, error = error)
 }